feat: add board endpoints

+++ server/src/services/board/deleteBoard.ts
@@ -5,9 +5,10 @@ import { connectToDB } from '../../utils/connectToDB';
 export const deleteBoardService = async (workSpaceId: string, boardId: string) => {
     if (!workSpaceId || !boardId) throw new Error('UserId or boardId not transferred').message
     await connectToDB()
+    const work = await WorkSpace.findOneAndUpdate<IWork>({_id: workSpaceId}, {$pull: {boards: boardId}})
+    if (!work) throw new Error('workId does not exist').message
     const board = await Board.findByIdAndDelete(boardId)
     if (!board) throw new Error('boardId does not exist').message
-    await WorkSpace.findOneAndUpdate<IWork>({_id: workSpaceId}, {$pull: {boards: boardId}})
 
     return board
 }
